fix: move init in authorizer itself
