+++ src/config/ess/configureESSEnvironment.ts
@@ -5,7 +5,6 @@ import { FeatureId } from '../../models/Feature';
 import { SettingsId } from '../../models/Settings';
 import { setTimezone, setDateTimeFormats } from '../setTimezoneAndFormat';
 import { Tokens } from '../Tokens';
-import { OpenIdClient } from './../../auth/OpenIdClient';
 
 async function setEssColourTheme() {
   const db = container.resolve<AdminDataSource>(Tokens.AdminDataSource);
@@ -85,30 +84,7 @@ async function enableDefaultEssFeatures() {
     ],
     true
   );
-  if (OpenIdClient.hasConfiguration()) {
-    const client = await OpenIdClient.getInstance();
-    const scopes = OpenIdClient.getScopes().join(' ');
 
-    const authUrl = client.authorizationUrl({ scope: scopes });
-
-    let endSessionUrl;
-    try {
-      endSessionUrl = client.endSessionUrl(); // try obtaining the end session url the standard way
-    } catch (e) {
-      endSessionUrl =
-        (client.issuer.ping_end_session_endpoint as string) ?? '/'; // try using PING ping_end_session_endpoint
-    }
-
-    await db.updateSettings({
-      settingsId: SettingsId.EXTERNAL_AUTH_LOGIN_URL,
-      settingsValue: authUrl,
-    });
-
-    await db.updateSettings({
-      settingsId: SettingsId.EXTERNAL_AUTH_LOGOUT_URL,
-      settingsValue: endSessionUrl,
-    });
-  }
   await db.updateSettings({
     settingsId: SettingsId.DEFAULT_INST_SCI_REVIEWER_FILTER,
     settingsValue: 'ME',
